switch back to u64

+++ src/model.rs
@@ -592,11 +592,11 @@ pub fn get_current_timestamp() -> u64 {
         .as_secs()
 }
 
-pub fn get_current_timestamp_ms() -> u128 {
+pub fn get_current_timestamp_ms() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .expect("Time went backwards")
-        .as_millis()
+        .as_millis() as u64
 }
 
 pub struct CommonRunContext {
